# src/test/resources/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "HRD Room Service API",
    "description" : "REST API for HRD room service operations, including authentication, academic years, classrooms, courses, and enrollments. Obtain a JWT from POST /api/v4/auth/login; send it as Bearer authentication on protected routes.",
    "contact" : {
      "name" : "HRD Room Service Team",
      "email" : "penhseyha4980@gmail.com"
    }
  },
  "paths" : {
...
    "/api/v4/courses/{courseId}" : {
      "get" : {
        "tags" : [ "course-controller" ],
        "operationId" : "byId",
        "parameters" : [ {
          "name" : "courseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCourseResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "course-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "courseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CourseUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCourseResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v4/enrollments/course/{courseId}" : {
      "get" : {
        "tags" : [ "enrollment-controller" ],
        "operationId" : "byCourse",
        "parameters" : [ {
          "name" : "courseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseEnrollmentResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v4/courses" : {
      "get" : {
        "tags" : [ "course-controller" ],
        "operationId" : "list_2",
        "parameters" : [ {
          "name" : "type",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "academicYearId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "includeArchived",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseCourseResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "EnrollmentResponse" : {
        "type" : "object",
        "properties" : {
          "enrollmentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "studentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "courseId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "courseName" : {
            "type" : "string"
          },
          "courseType" : {
            "type" : "string"
          },
          "academicYearId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "academicYearName" : {
            "type" : "string"
          },
          "generation" : {
            "type" : "integer",
            "format" : "int32"
          },
          "enrolledAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isPassed" : {
            "type" : "boolean"
          },
          "isTerminated" : {
            "type" : "boolean"
          },
          "terminatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "terminationReason" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "CourseUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "description" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
      "ApiResponseCourseResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/CourseResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "CourseResponse" : {
        "type" : "object",
        "properties" : {
          "courseId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "code" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "academicYearId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "prerequisiteCourseId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "description" : {
            "type" : "string"
          },
          "isArchived" : {
            "type" : "boolean"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponsePageResponseEnrollmentResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseEnrollmentResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "PageResponseEnrollmentResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EnrollmentResponse"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ApiResponsePageResponseCourseResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseCourseResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "PageResponseCourseResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CourseResponse"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}